requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
idna==3.2
Pillow==8.3.2
py3minepi==0.0.1
urllib3==1.26.7
